// src/app/services/projects/models/create-task-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CreateTaskRequest {
  assigneeId?: string;
  description?: string;
  dueDate?: string;
  priority?: 'LOW' | 'MEDIUM' | 'HIGH' | 'CRITICAL';
  projectId: string;
  sprintId?: string;
  storyPoints?: number;
  taskType?: 'TASK' | 'BUG' | 'STORY' | 'EPIC' | 'SUBTASK';
  title: string;
}